+++ threedi_schema/tests/test_gpkg.py
@@ -1,47 +1,31 @@
-import re
-import subprocess
-
 import pytest
+from sqlalchemy import text
 
-from threedi_schema.application.schema import get_schema_version
 
+@pytest.mark.parametrize("upgrade_spatialite", [True, False])
+def test_convert_to_geopackage(oldest_sqlite, upgrade_spatialite):
+    if upgrade_spatialite:
+        oldest_sqlite.schema.upgrade(upgrade_spatialite_version=True)
 
-def test_convert_to_geopackage(oldest_sqlite):
-    if get_schema_version() < 300:
-        pytest.skip("Gpkg not supported for schema < 300")
-    # In case the fixture changes and refers to a geopackage,
-    # convert_to_geopackage will be ignored because the db is already a geopackage
-    assert oldest_sqlite.get_engine().dialect.name == "sqlite"
-    # check if ogr2ogr is installed and has the right version:
-    result = subprocess.run(
-        "ogr2ogr --version",
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True,
-    )
-    expect_success = False
-    if result.returncode == 0:
-        # get version
-        version = re.findall(r"\b(\d+\.\d+\.\d+)\b", result.stdout)[0]
-        # trim patch version and convert to float
-        float_version = float(version[0 : version.rfind(".")])
-        if float_version >= 3.4:
-            expect_success = True
-    if expect_success:
-        oldest_sqlite.schema.convert_to_geopackage()
-        # Ensure that after the conversion the geopackage is used
-        assert oldest_sqlite.path.suffix == ".gpkg"
-        assert oldest_sqlite.get_engine().dialect.name == "geopackage"
-        assert oldest_sqlite.schema.validate_schema()
-    else:
-        # Upgrade is not performed in convert_to_geopackage when ogr2ogr doesn't run
-        # because no operation should be performed in that case.
-        # However, this wil result in an invalid schema, so here we will run upgrade manually.
-        # Because convert_to_geopackage() is only used via upgrade; this is not an issue.
-        with pytest.warns():
-            oldest_sqlite.schema.convert_to_geopackage()
-            oldest_sqlite.schema.upgrade()
-            assert oldest_sqlite.path.suffix == ".sqlite"
-            assert oldest_sqlite.get_engine().dialect.name == "sqlite"
-            assert oldest_sqlite.schema.validate_schema()
+    oldest_sqlite.schema.convert_to_geopackage()
+    # Ensure that after the conversion the geopackage is used
+    assert oldest_sqlite.path.suffix == ".gpkg"
+    with oldest_sqlite.session_scope() as session:
+        gpkg_table_exists = bool(
+            session.execute(
+                text(
+                    "SELECT count(*) FROM sqlite_master WHERE type='table' AND name='gpkg_contents';"
+                )
+            ).scalar()
+        )
+        spatialite_table_exists = bool(
+            session.execute(
+                text(
+                    "SELECT count(*) FROM sqlite_master WHERE type='table' AND name='spatial_ref_sys';"
+                )
+            ).scalar()
+        )
+
+    assert gpkg_table_exists
+    assert not spatialite_table_exists
+    assert oldest_sqlite.schema.validate_schema()